+++ recipes/install_prereq_packages.rb
@@ -1,15 +1,14 @@
+include_recipe 'build-essential'
+
 log "\n=========== Start MapR install_mapr_prereqs.rb =============\n"
 
 package 'bash'
 package 'rpcbind'
 package 'dmidecode'
-package 'glibc'
 package 'hdparm'
 package 'initscripts'
 package 'iputils'
 package 'irqbalance'
-package 'libgcc'
-package 'libstdc++'
 package 'redhat-lsb-core'
 package 'rpm-libs'
 package 'sdparm'
@@ -23,34 +22,32 @@ package 'git'
 package 'nfs-utils'
 package 'nfs-utils-lib'
 package 'git'
-package 'gcc'
 package 'patch'
 package 'dstat'
 package 'lsof'
 
-java_version= "#{node['java']['version']}"
+include_recipe 'java' if node['mapr']['manage_java']
 
+# java_version = node['java']['version']
 
-bash 'Install_java' do
-  code <<-EOH
-    yum -y install #{java_version} 
-  EOH
-end
+# bash 'Install_java' do
+#   code <<-EOH
+#     yum -y install #{java_version}
+#   EOH
+# end
 
-#Add JAVA_HOME to /etc/profile
-ruby_block "Set JAVA_HOME in /etc/profile" do
-  block do
-        file  = Chef::Util::FileEdit.new("/etc/profile")
-        file.insert_line_if_no_match("export JAVA_HOME=#{node[:java][:home]}","\nexport JAVA_HOME=#{node[:java][:home]}")
-	file.insert_line_if_no_match("export EDITOR=vi","export EDITOR=vi")
+# # Add JAVA_HOME to /etc/profile
+# ruby_block 'Set JAVA_HOME in /etc/profile' do
+#   block do
+#     file = Chef::Util::FileEdit.new('/etc/profile')
+#     file.insert_line_if_no_match("export JAVA_HOME=#{node['java']['home']}",
+#                                  "\nexport JAVA_HOME=#{node['java']['home']}")
+#     file.insert_line_if_no_match('export EDITOR=vi', 'export EDITOR=vi')
 
-	file.write_file
-  end
-end
+#     file.write_file
+#   end
+# end
 
-bash 'turn_on_rpcbind' do
-  code <<-EOH
-    service rpcbind start
-    chkconfig rpcbind on
-  EOH
+service 'rpcbind' do
+  action [:enable, :start]
 end