+++ recipes/iptables.rb
@@ -1,6 +1,3 @@
 log "\n=========== Start MapR iptables.rb =============\n"
 
-service 'iptables' do
-	action [:stop, :disable]
-end
-
+include_recipe 'iptables::disabled'