+++ recipes/user_mapr.rb
@@ -1,36 +1,19 @@
 log "\n=========== Start MapR user_mapr.rb =============\n"
 
-group node['mapr']['group'] do
-  gid node['mapr']['gid']
+data_bag   = node['mapr']['users_data_bag']
+mapr_group = node['mapr']['group']
+mapr_gid   = node['mapr']['gid']
+
+users_manage mapr_group do
+  group_id mapr_gid
+  action [:create]
+  data_bag data_bag
+  manage_nfs_home_dirs false
+  not_if { data_bag.nil? }
 end
 
-user node['mapr']['user'] do
-  uid node['mapr']['uid']
-  gid node['mapr']['gid']
-  shell "/bin/bash"
-  home "/home/#{node['mapr']['user']}"
+sudo 'mapr' do
+  user 'mapr'
+  runas 'ALL'
+  nopasswd true
 end
-
-user "setting mapr password" do
-  username "#{node['mapr']['user']}"
-  password "#{node['mapr']['password']}"
-  action :modify
-end
-
-
-directory "/home/#{node['mapr']['user']}" do
-  owner node['mapr']['user']
-  group node['mapr']['group']
-  mode 0700
-end
-
-ruby_block "Add mapr to /etc/sudoers" do
-  block do
-	file  = Chef::Util::FileEdit.new("/etc/sudoers")
-        file.insert_line_after_match("root    ALL=(ALL)       ALL","mapr	ALL=(ALL) 	ALL")
-	file.insert_line_if_no_match("mapr      ALL=(ALL)       ALL", "mapr      ALL=(ALL)       ALL")
-
-        file.write_file
-  end
-end
-