+++ recipes/validate_host.rb
@@ -1,73 +1,58 @@
 log "\n=========== Start MapR validate_host.rb =============\n"
 
-bash "uname -m" do
-	code <<-EOF
-		 echo `uname -m`
-	EOF
+gpgkey_url = node['mapr']['yum']['gpgkey_url']
+
+include_recipe 'sysctl::default'
+
+bash 'uname -m' do
+  code <<-EOF
+    uname -m
+  EOF
 end
 
 execute 'validate_host_viable' do
-	command "echo `uname -m`"
-	action :run
+  command 'uname -m'
+  action :run
 end
 
-ruby_block "edit etc sysctl" do
-  block do
-	file  = Chef::Util::FileEdit.new("/etc/sysctl.conf")
-	file.search_file_delete_line ("#MapR Values")
-	file.insert_line_if_no_match("#MapR\ Values", "\n#MapR Values")
-	
-        file.search_file_delete_line ("vm.swappiness")
-	file.insert_line_if_no_match("vm.swappiness =","vm.swappiness = 0")
+sysctl_param 'vm.swappiness' do
+  value 0
+end
 
-	file.search_file_delete_line ("net.ipv4.tcp_retries2")
-	file.insert_line_if_no_match("net.ipv4.tcp_retries2","net.ipv4.tcp_retries2 = 5")
+sysctl_param 'net.ipv4.tcp_retries2' do
+  value 5
+end
 
-	file.search_file_delete_line ("vm.overcommit_memory") 
-	file.insert_line_if_no_match("vm.overcommit_memory","vm.overcommit_memory = 0 \n")
-	
-	file.write_file	
+sysctl_param 'vm.overcommit_memory' do
+  value 0
+end
+
+%w( hard soft ).each do |ltype|
+  set_limit 'mapr' do
+    type ltype
+    item 'nofile'
+    value 64_000
+  end
+
+  set_limit 'mapr' do
+    type ltype
+    item 'nproc'
+    value 64_000
   end
 end
 
-ruby_block "Edit /etc/security/limits.conf" do
-  block do
-        file  = Chef::Util::FileEdit.new("/etc/security/limits.conf")
-	file.search_file_delete_line ("mapr")
-	file.search_file_delete_line ("#End of")
-        file.insert_line_if_no_match("mapr","mapr	-	nofile	64000")
-	file.insert_line_if_no_match("#End of","#End of file")	
-	file.write_file
-  end
-end
-	
-ruby_block "Edit /etc/security/limits.d/90-nproc.conf" do
-  block do
-	file  = Chef::Util::FileEdit.new("/etc/security/limits.d/90-nproc.conf")
-        file.search_file_delete_line ("mapr")
-        file.search_file_delete_line ("#End of")
-        file.insert_line_if_no_match("mapr","mapr	-	nproc	64000")
-        file.insert_line_if_no_match("#End of","#End of file")
-
-        file.write_file
-  end
-end	
-
-
-file '/etc/yum.repos.d/maprtech.repo' do
-content "[maprtech]
-name=MapR Technologies
-baseurl=http://package.mapr.com/releases/v#{node['mapr']['version']}/redhat/
-enabled=1
-gpgcheck=0
-protect=1
- 
-[maprecosystem]
-name=MapR Technologies
-baseurl=http://package.mapr.com/releases/ecosystem-4.x/redhat
-enabled=1
-gpgcheck=0
-protect=1"
-
+yum_repository 'maprtech' do
+  description 'MapR Technologies'
+  baseurl "http://package.mapr.com/releases/v#{node['mapr']['version']}/redhat"
+  gpgcheck true
+  gpgkey gpgkey_url
+  action :create
 end
 
+yum_repository 'maprecosystem' do
+  description 'MapR Technologies (ecosystem)'
+  baseurl 'http://package.mapr.com/releases/ecosystem-4.x/redhat'
+  gpgcheck false
+  gpgkey gpgkey_url
+  action :create
+end