+++ CHANGELOG.md
@@ -3,6 +3,43 @@ mapr_installation CHANGELOG
 
 This file is used to list changes made in each version of the mapr_installation cookbook.
 
+## v0.6.2
+
+* set the core-site.xml action and notifies
+
+## v0.5.0
+
+* fix bug in ::spark_master
+## v0.5.0
+
+* install drill
+
+## v0.4.0
+
+* install Apache Spark
+
+## v0.3.2
+
+* add the rpm gpgkey
+* turn more ruby blocks into bash blocks
+
+## v0.3.1
+
+* make the /etc/clustershell directory
+* this version successfully launched a single node cluster
+
+## v0.3.0
+
+* switch to java cookbook
+* some fixes
+
+## v0.2.1
+
+* don't blast root's ssh keys and config
+
+
+## v0.2.0
+
 0.1.0
 -----
 - [your_name] - Initial release of mapr_installation